Prepare revocation of ephemeral tokens
If the plugin logs into Vault in the process of executing a task beyond
just logging in (i.e. not in e.g. the token_create lookup), the tokens
should be revoked at the end of the action to prevent them from leaking.

In order to support that, the authentication method needs to tell us
whether it created a fresh token (which we should revoke) or whether it
used a token provided by the caller (which we should not revoke).